test: Customer 인수 테스트, Controller 테스트 분리
