+++ src/test/java/woowacourse/shoppingcart/acceptance/CustomerAcceptanceTest.java
@@ -5,9 +5,6 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.NullSource;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
@@ -16,7 +13,6 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import woowacourse.auth.dto.LoginRequest;
 import woowacourse.auth.dto.LoginTokenResponse;
-import woowacourse.shoppingcart.dto.ErrorResponseWithField;
 import woowacourse.shoppingcart.dto.customer.CustomerCreateRequest;
 import woowacourse.shoppingcart.dto.customer.CustomerResponse;
 import woowacourse.shoppingcart.dto.customer.CustomerUpdateRequest;
@@ -34,78 +30,6 @@ public class CustomerAcceptanceTest extends AcceptanceTest {
             () -> assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value()),
             () -> assertThat(response.header("Location")).startsWith("/api/customers/")
         );
-    }
-
-    @DisplayName("회원가입시 중복된 email로 가입하려는 경우 400 응답을 반환한다.")
-    @Test
-    void addCustomer_duplicated_email() {
-        ExtractableResponse<Response> response = 회원가입_요청(
-            new CustomerCreateRequest("puterism@naver.com", "roma", "12345678"));
-
-        ErrorResponseWithField errorResponseWithField = response.as(ErrorResponseWithField.class);
-
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(errorResponseWithField).usingRecursiveComparison()
-            .isEqualTo(new ErrorResponseWithField("email", "이미 가입된 이메일입니다."));
-    }
-
-    @DisplayName("회원가입시 중복된 username으로 가입하려는 경우 400 응답을 반환한다.")
-    @Test
-    void addCustomer_duplicated_username() {
-        ExtractableResponse<Response> response = 회원가입_요청(
-            new CustomerCreateRequest("philz@naver.com", "puterism", "12345678"));
-
-        ErrorResponseWithField errorResponseWithField = response.as(ErrorResponseWithField.class);
-
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(errorResponseWithField).usingRecursiveComparison()
-            .isEqualTo(new ErrorResponseWithField("username", "이미 가입된 닉네임입니다."));
-    }
-
-    @DisplayName("회원 가입시 잘못된 형식의 email을 입력한 경우 400 응답을 반환한다.")
-    @ParameterizedTest
-    @ValueSource(strings = {"not_email_format", "philz @gmail.com", ""})
-    @NullSource
-    void create_exception_parameter_email(String email) {
-        // given
-        CustomerCreateRequest customerCreateRequest = new CustomerCreateRequest(email, "philz", "12345678");
-
-        // when
-        ExtractableResponse<Response> response = 회원가입_요청(customerCreateRequest);
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    @DisplayName("회원 가입시 잘못된 형식의 username을 입력한 경우 400 응답을 반환한다.")
-    @ParameterizedTest
-    @ValueSource(strings = {"01234567890", "", " "})
-    @NullSource
-    void create_exception_parameter_name(String username) {
-        // given
-        CustomerCreateRequest customerCreateRequest = new CustomerCreateRequest("philz@gmail.com", username,
-            "12345678");
-
-        // when
-        ExtractableResponse<Response> response = 회원가입_요청(customerCreateRequest);
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    @DisplayName("회원 가입시 잘못된 형식의 password을 입력한 경우 400 응답을 반환한다.")
-    @ParameterizedTest
-    @ValueSource(strings = {"", " ", "1234", "012345678901234567890"})
-    @NullSource
-    void create_exception_parameter_password(String password) {
-        // given
-        CustomerCreateRequest customerCreateRequest = new CustomerCreateRequest("philz@gmail.com", "philz", password);
-
-        // when
-        ExtractableResponse<Response> response = 회원가입_요청(customerCreateRequest);
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
     }
 
     @DisplayName("내 정보 조회")
@@ -127,23 +51,6 @@ public class CustomerAcceptanceTest extends AcceptanceTest {
         );
     }
 
-    @DisplayName("다른 사람의 정보 조회를 하면 403을 반환한다")
-    @Test
-    void findCustomer_otherId() {
-        // given
-        ExtractableResponse<Response> createResponse = 회원가입_요청(
-            new CustomerCreateRequest("roma@naver.com", "roma", "12345678"));
-        long savedId = ID_추출(createResponse);
-        long 다른사람의_ID = savedId + 1L;
-        String token = 로그인_요청_및_토큰발급(new LoginRequest("roma@naver.com", "12345678"));
-
-        // when
-        ExtractableResponse<Response> response = 회원조회_요청(token, 다른사람의_ID);
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.FORBIDDEN.value());
-    }
-
     @DisplayName("내 정보 수정")
     @Test
     void update() {
@@ -161,68 +68,6 @@ public class CustomerAcceptanceTest extends AcceptanceTest {
         );
     }
 
-    @DisplayName("내 정보 수정시 기존과 같은 username을 입력한 경우 200 OK를 응답한다.")
-    @Test
-    void update_same_origin_name() {
-        long savedId = 회원가입_요청_및_ID_추출(new CustomerCreateRequest("roma@naver.com", "roma", "12345678"));
-
-        String token = 로그인_요청_및_토큰발급(new LoginRequest("roma@naver.com", "12345678"));
-        ExtractableResponse<Response> response = 회원정보수정_요청(token, savedId, new CustomerUpdateRequest("roma"));
-
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    @DisplayName("내 정보 수정 시 중복된 username으로 변경하려는 경우 400 응답을 반환한다.")
-    @Test
-    void update_duplicated_username() {
-        long savedId = 회원가입_요청_및_ID_추출(new CustomerCreateRequest("roma@naver.com", "roma", "12345678"));
-
-        String token = 로그인_요청_및_토큰발급(new LoginRequest("roma@naver.com", "12345678"));
-        ExtractableResponse<Response> response = 회원정보수정_요청(token, savedId, new CustomerUpdateRequest("yujo11"));
-
-        ErrorResponseWithField errorResponseWithField = response.as(ErrorResponseWithField.class);
-
-        assertAll(
-            () -> assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value()),
-            () -> assertThat(errorResponseWithField).usingRecursiveComparison()
-                .isEqualTo(new ErrorResponseWithField("username", "이미 가입된 닉네임입니다."))
-        );
-    }
-
-    @DisplayName("내 정보 수정시 잘못된 형식의 username을 입력한 경우 400 응답을 반환한다.")
-    @ParameterizedTest
-    @ValueSource(strings = {"01234567890", "", " "})
-    @NullSource
-    void update_exception_parameter_name(String username) {
-        // given
-        String token = 로그인_요청_및_토큰발급(new LoginRequest("puterism@naver.com", "12349053145"));
-        CustomerUpdateRequest customerUpdateRequest = new CustomerUpdateRequest(username);
-
-        // when
-        ExtractableResponse<Response> response = 회원정보수정_요청(token, 1L, customerUpdateRequest);
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    @DisplayName("다른 사람의 정보를 수정하면 403을 반환한다")
-    @Test
-    void update_otherId() {
-        // given
-        ExtractableResponse<Response> createResponse = 회원가입_요청(
-            new CustomerCreateRequest("roma@naver.com", "roma", "12345678"));
-        long savedId = ID_추출(createResponse);
-        long 다른사람의_ID = savedId + 1L;
-        String token = 로그인_요청_및_토큰발급(new LoginRequest("roma@naver.com", "12345678"));
-
-        // when
-        CustomerUpdateRequest updateRequest = new CustomerUpdateRequest("philz");
-        ExtractableResponse<Response> response = 회원정보수정_요청(token, 다른사람의_ID, updateRequest);
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.FORBIDDEN.value());
-    }
-
     @DisplayName("회원탈퇴")
     @Test
     void delete() {
@@ -232,17 +77,7 @@ public class CustomerAcceptanceTest extends AcceptanceTest {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
     }
 
-    @DisplayName("다른 사람의 정보를 수정하면 403을 반환한다")
-    @Test
-    void delete_otherId() {
-        String token = 로그인_요청_및_토큰발급(new LoginRequest("puterism@naver.com", "12349053145"));
-        long 다른사람의_ID = 2L;
-        ExtractableResponse<Response> response = 회원탈퇴_요청(token, 다른사람의_ID);
-
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.FORBIDDEN.value());
-    }
-
-    public String 로그인_요청_및_토큰발급(LoginRequest request) {
+    private String 로그인_요청_및_토큰발급(LoginRequest request) {
         ExtractableResponse<Response> loginResponse = RestAssured
             .given().log().all()
             .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -255,7 +90,7 @@ public class CustomerAcceptanceTest extends AcceptanceTest {
         return loginTokenResponse.getAccessToken();
     }
 
-    public ExtractableResponse<Response> 회원가입_요청(CustomerCreateRequest request) {
+    private ExtractableResponse<Response> 회원가입_요청(CustomerCreateRequest request) {
         return RestAssured
             .given().log().all()
             .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -265,11 +100,11 @@ public class CustomerAcceptanceTest extends AcceptanceTest {
             .extract();
     }
 
-    public long 회원가입_요청_및_ID_추출(CustomerCreateRequest request) {
+    private long 회원가입_요청_및_ID_추출(CustomerCreateRequest request) {
         return ID_추출(회원가입_요청(request));
     }
 
-    public ExtractableResponse<Response> 회원조회_요청(String token, Long id) {
+    private ExtractableResponse<Response> 회원조회_요청(String token, Long id) {
         return RestAssured
             .given().log().all()
             .header("Authorization", "Bearer " + token)
@@ -278,7 +113,7 @@ public class CustomerAcceptanceTest extends AcceptanceTest {
             .extract();
     }
 
-    public ExtractableResponse<Response> 회원정보수정_요청(String token, long id, CustomerUpdateRequest request) {
+    private ExtractableResponse<Response> 회원정보수정_요청(String token, long id, CustomerUpdateRequest request) {
         return RestAssured
             .given().log().all()
             .header("Authorization", "Bearer " + token)
@@ -289,7 +124,7 @@ public class CustomerAcceptanceTest extends AcceptanceTest {
             .extract();
     }
 
-    public ExtractableResponse<Response> 회원탈퇴_요청(String token, long id) {
+    private ExtractableResponse<Response> 회원탈퇴_요청(String token, long id) {
         return RestAssured
             .given().log().all()
             .header("Authorization", "Bearer " + token)
